1. Rework on Period.java 2. Bug fix on Date#sub() 3. Some code review on TimeGrid

+++ jquantlib/src/main/java/org/jquantlib/methods/lattices/Joshi4.java
@@ -62,7 +62,7 @@ public class Joshi4 extends BinomialTree {
 
 	@Override
     public double underlying(final int i, final int index) {
-        final long j = (long) i - (long) index;
+        final int j = i - index;
         final double d = j;
 		return x0 * Math.pow(down, d) * Math.pow(up, index);
